Convert i32 to i64 and clippy code

+++ algonaut_indexer/src/models/account.rs
@@ -17,10 +17,10 @@ pub struct Account {
     pub address: String,
     /// \\[algo\\] total number of MicroAlgos in the account
     #[serde(rename = "amount")]
-    pub amount: i32,
+    pub amount: u64,
     /// specifies the amount of MicroAlgos in the account, without the pending rewards.
     #[serde(rename = "amount-without-pending-rewards")]
-    pub amount_without_pending_rewards: i32,
+    pub amount_without_pending_rewards: u64,
     /// \\[appl\\] applications local data stored in this account.  Note the raw object uses `map[int] -> AppLocalState` for this type.
     #[serde(rename = "apps-local-state", skip_serializing_if = "Option::is_none")]
     pub apps_local_state: Option<Vec<crate::models::ApplicationLocalState>>,
@@ -29,7 +29,7 @@ pub struct Account {
         rename = "apps-total-extra-pages",
         skip_serializing_if = "Option::is_none"
     )]
-    pub apps_total_extra_pages: Option<i32>,
+    pub apps_total_extra_pages: Option<u64>,
     #[serde(rename = "apps-total-schema", skip_serializing_if = "Option::is_none")]
     pub apps_total_schema: Option<Box<crate::models::ApplicationStateSchema>>,
     /// \\[asset\\] assets held by this account.  Note the raw object uses `map[int] -> AssetHolding` for this type.
@@ -40,7 +40,7 @@ pub struct Account {
     pub auth_addr: Option<String>,
     /// Round during which this account was most recently closed.
     #[serde(rename = "closed-at-round", skip_serializing_if = "Option::is_none")]
-    pub closed_at_round: Option<i32>,
+    pub closed_at_round: Option<u64>,
     /// \\[appp\\] parameters of applications created by this account including app global data.  Note: the raw account uses `map[int] -> AppParams` for this type.
     #[serde(rename = "created-apps", skip_serializing_if = "Option::is_none")]
     pub created_apps: Option<Vec<crate::models::Application>>,
@@ -49,7 +49,7 @@ pub struct Account {
     pub created_assets: Option<Vec<crate::models::Asset>>,
     /// Round during which this account first appeared in a transaction.
     #[serde(rename = "created-at-round", skip_serializing_if = "Option::is_none")]
-    pub created_at_round: Option<i32>,
+    pub created_at_round: Option<u64>,
     /// Whether or not this account is currently closed.
     #[serde(rename = "deleted", skip_serializing_if = "Option::is_none")]
     pub deleted: Option<bool>,
@@ -57,16 +57,16 @@ pub struct Account {
     pub participation: Option<Box<crate::models::AccountParticipation>>,
     /// amount of MicroAlgos of pending rewards in this account.
     #[serde(rename = "pending-rewards")]
-    pub pending_rewards: i32,
+    pub pending_rewards: u64,
     /// \\[ebase\\] used as part of the rewards computation. Only applicable to accounts which are participating.
     #[serde(rename = "reward-base", skip_serializing_if = "Option::is_none")]
-    pub reward_base: Option<i32>,
+    pub reward_base: Option<u64>,
     /// \\[ern\\] total rewards of MicroAlgos the account has received, including pending rewards.
     #[serde(rename = "rewards")]
-    pub rewards: i32,
+    pub rewards: u64,
     /// The round for which this information is relevant.
     #[serde(rename = "round")]
-    pub round: i32,
+    pub round: u64,
     /// Indicates what type of signature is used by this account, must be one of: * sig * msig * lsig * or null if unknown
     #[serde(rename = "sig-type", skip_serializing_if = "Option::is_none")]
     pub sig_type: Option<SigType>,
@@ -75,40 +75,41 @@ pub struct Account {
     pub status: String,
     /// The count of all applications that have been opted in, equivalent to the count of application local data (AppLocalState objects) stored in this account.
     #[serde(rename = "total-apps-opted-in")]
-    pub total_apps_opted_in: i32,
+    pub total_apps_opted_in: u64,
     /// The count of all assets that have been opted in, equivalent to the count of AssetHolding objects held by this account.
     #[serde(rename = "total-assets-opted-in")]
-    pub total_assets_opted_in: i32,
+    pub total_assets_opted_in: u64,
     /// For app-accounts only. The total number of bytes allocated for the keys and values of boxes which belong to the associated application.
     #[serde(rename = "total-box-bytes")]
-    pub total_box_bytes: i32,
+    pub total_box_bytes: u64,
     /// For app-accounts only. The total number of boxes which belong to the associated application.
     #[serde(rename = "total-boxes")]
-    pub total_boxes: i32,
+    pub total_boxes: u64,
     /// The count of all apps (AppParams objects) created by this account.
     #[serde(rename = "total-created-apps")]
-    pub total_created_apps: i32,
+    pub total_created_apps: u64,
     /// The count of all assets (AssetParams objects) created by this account.
     #[serde(rename = "total-created-assets")]
-    pub total_created_assets: i32,
+    pub total_created_assets: u64,
 }
 
 impl Account {
     /// Account information at a given round.  Definition: data/basics/userBalance.go : AccountData
+    #[allow(clippy::too_many_arguments)]
     pub fn new(
         address: String,
-        amount: i32,
-        amount_without_pending_rewards: i32,
-        pending_rewards: i32,
-        rewards: i32,
-        round: i32,
+        amount: u64,
+        amount_without_pending_rewards: u64,
+        pending_rewards: u64,
+        rewards: u64,
+        round: u64,
         status: String,
-        total_apps_opted_in: i32,
-        total_assets_opted_in: i32,
-        total_box_bytes: i32,
-        total_boxes: i32,
-        total_created_apps: i32,
-        total_created_assets: i32,
+        total_apps_opted_in: u64,
+        total_assets_opted_in: u64,
+        total_box_bytes: u64,
+        total_boxes: u64,
+        total_created_apps: u64,
+        total_created_assets: u64,
     ) -> Account {
         Account {
             address,
