+++ algonaut_indexer/src/models/account_participation.rs
@@ -20,13 +20,13 @@ pub struct AccountParticipation {
     pub state_proof_key: Option<String>,
     /// \\[voteFst\\] First round for which this participation is valid.
     #[serde(rename = "vote-first-valid")]
-    pub vote_first_valid: i32,
+    pub vote_first_valid: u64,
     /// \\[voteKD\\] Number of subkeys in each batch of participation keys.
     #[serde(rename = "vote-key-dilution")]
-    pub vote_key_dilution: i32,
+    pub vote_key_dilution: u64,
     /// \\[voteLst\\] Last round for which this participation is valid.
     #[serde(rename = "vote-last-valid")]
-    pub vote_last_valid: i32,
+    pub vote_last_valid: u64,
     /// \\[vote\\] root participation public key (if any) currently registered for this round.
     #[serde(rename = "vote-participation-key")]
     pub vote_participation_key: String,
@@ -36,9 +36,9 @@ impl AccountParticipation {
     /// AccountParticipation describes the parameters used by this account in consensus protocol.
     pub fn new(
         selection_participation_key: String,
-        vote_first_valid: i32,
-        vote_key_dilution: i32,
-        vote_last_valid: i32,
+        vote_first_valid: u64,
+        vote_key_dilution: u64,
+        vote_last_valid: u64,
         vote_participation_key: String,
     ) -> AccountParticipation {
         AccountParticipation {
