+++ algonaut_indexer/src/models/application_local_state.rs
@@ -17,26 +17,26 @@ pub struct ApplicationLocalState {
         rename = "closed-out-at-round",
         skip_serializing_if = "Option::is_none"
     )]
-    pub closed_out_at_round: Option<i32>,
+    pub closed_out_at_round: Option<u64>,
     /// Whether or not the application local state is currently deleted from its account.
     #[serde(rename = "deleted", skip_serializing_if = "Option::is_none")]
     pub deleted: Option<bool>,
     /// The application which this local state is for.
     #[serde(rename = "id")]
-    pub id: i32,
+    pub id: u64,
     /// Represents a key-value store for use in an application.
     #[serde(rename = "key-value", skip_serializing_if = "Option::is_none")]
     pub key_value: Option<Vec<crate::models::TealKeyValue>>,
     /// Round when the account opted into the application.
     #[serde(rename = "opted-in-at-round", skip_serializing_if = "Option::is_none")]
-    pub opted_in_at_round: Option<i32>,
+    pub opted_in_at_round: Option<u64>,
     #[serde(rename = "schema")]
     pub schema: Box<crate::models::ApplicationStateSchema>,
 }
 
 impl ApplicationLocalState {
     /// Stores local state associated with an application.
-    pub fn new(id: i32, schema: crate::models::ApplicationStateSchema) -> ApplicationLocalState {
+    pub fn new(id: u64, schema: crate::models::ApplicationStateSchema) -> ApplicationLocalState {
         ApplicationLocalState {
             closed_out_at_round: None,
             deleted: None,
