+++ algonaut_indexer/src/models/lookup_account_created_assets_200_response.rs
@@ -14,7 +14,7 @@ pub struct LookupAccountCreatedAssets200Response {
     pub assets: Vec<crate::models::Asset>,
     /// Round at which the results were computed.
     #[serde(rename = "current-round")]
-    pub current_round: i32,
+    pub current_round: u64,
     /// Used for pagination, when making another request provide this token with the next parameter.
     #[serde(rename = "next-token", skip_serializing_if = "Option::is_none")]
     pub next_token: Option<String>,
@@ -23,7 +23,7 @@ pub struct LookupAccountCreatedAssets200Response {
 impl LookupAccountCreatedAssets200Response {
     pub fn new(
         assets: Vec<crate::models::Asset>,
-        current_round: i32,
+        current_round: u64,
     ) -> LookupAccountCreatedAssets200Response {
         LookupAccountCreatedAssets200Response {
             assets,
